src/test/java/me/toyproject/springboottodoexample/domain/account/AccountServiceTest.java
package me.toyproject.springboottodoexample.domain.account;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.Set;

import static org.assertj.core.api.Assertions.assertThat;

@RunWith(SpringRunner.class)
@SpringBootTest
@ActiveProfiles("test")
public class AccountServiceTest {

    @Autowired
    AccountService accountService;

    @Autowired
    PasswordEncoder passwordEncoder;

    @Test
    public void findByUserName() {

        String name = "sunkyung";
        String password = "test";

        //Given
        Account account = Account.builder()
                .name(name)
                .password(passwordEncoder.encode(password))
                .roles(Set.of(AccountRole.ADMIN, AccountRole.USER))
                .build();
        this.accountService.saveAccount(account);

        //When
        UserDetailsService userDetailsService = accountService;
        UserDetails userDetails = userDetailsService.loadUserByUsername(name);

        //Then
        assertThat(userDetails.getPassword()).isEqualTo(account.getPassword());
    }

    @Test(expected = UsernameNotFoundException.class)
    public void findByUsernameFail() {
        String failName = "failName";
        accountService.loadUserByUsername(failName);
    }
}